requirements.txt
torch
torchvision
pandas
captum
scipy
numpy
scikit-learn
geopandas
shapely
matplotlib
opencv-python-headless
streamlit
plotly
fastapi
uvicorn 
python-multipart
python-dotenv
gunicorn
